Move default BasicMenuBuilder#rowPrefixSupplier to RowPrefixSupplier
